fix: moved authOptions to own file to correct compile fail on veercel
